// src/modules/app/components/routes/types.ts
export enum Paths {
  root = '/',
  vaults = '/vaults',
  pools = '/pools',
  swap = '/swap',
  liquidity = '/liquidity',
  planetToken = '/planet-token',
  lending = '/lending',
  liquidations = '/liquidations',
  markets = '/markets',
  metrics = '/metrics',
  pink = '/pink',
  pinkExplore = '/pink/explore',
  pinkProfileview = '/pink/profile',
  pinkProfileedit = '/pink/profile/edit',
  pinkcreatenftsingle = '/pink/create/single',
  pinkcreatenftmultiple = '/pink/create/multiple',
  pinkcreatenfttype = '/pink/create/type',
  pinkasset = '/pink/asset',
  pinkfaq = '/pink/faq',
  pinkactivity = '/pink/activity',
  pinkcollection = '/pink/collection',
  collectionedit = '/pink/collection/edit',
  
}